pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "freemobile-classifier"
version = "2.0.0"
description = "Classification automatique de réclamations avec Gemini API et multi-modèles"
authors = [{name = "FreeMobilaChat Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "streamlit>=1.32.0",
    "google-generativeai>=0.4.1",
    "pydantic>=2.6.3",
    "pandas>=2.2.1",
    "plotly>=5.19.0",
    "numpy>=1.26.4",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "tqdm>=4.65.0",
    "psutil>=5.9.5",
    "tenacity>=8.2.3",
    "openpyxl>=3.1.2",
    "emoji>=2.8.0",
    "unidecode>=1.3.0",
    "chardet>=5.2.0",
    "bcrypt>=4.0.1",
    "pyjwt>=2.8.0",
    "sqlalchemy>=2.0.21",
    "loguru>=0.7.2",
    "anthropic>=0.8.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.5",
    "pytest-mock>=3.12.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pylint>=3.1.0",
    "isort>=5.12.0",
    "ipython>=8.17.2",
    "jupyter>=1.0.0",
]

test = [
    "pytest>=8.1.1",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.5",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
    "playwright>=1.42.0",
    "pytest-playwright>=0.4.4",
    "responses>=0.25.0",
    "faker>=24.0.0",
    "scikit-learn>=1.4.1",
    "seaborn>=0.13.2",
]

ml = [
    "transformers>=4.34.0",
    "torch>=2.1.0",
    "sentencepiece>=0.1.99",
    "ollama>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.isort]
profile = "black"
line_length = 100



